refactor: Move the hlist in a stand alone module
