+++ src/lib.rs
@@ -3,4 +3,5 @@
 //! Provides all the building blocks to serve [Web Of Things](https://www.w3.org/WoT/) Things.
 
 pub mod advertise;
+pub mod hlist;
 pub mod servient;
